Move scheduling functionality to dedicated subdirectory
This is a step towards implementing a regular cleanup job for suite
working directories.

CMK-14849
